refactor: timeline components
