+++ web/src/lib/components/timeline/Scrubber.svelte
